requirements.txt
Django==4.2
djangorestframework==3.14
Pillow>=10.0
django-cors-headers==3.14
whitenoise==6.6
